+++ READMEZH.md
@@ -1,0 +1,5 @@
+In my initial implementation, I utilized the ChatGPT API to process user input combined with a predefined prompt to identify the user intent. This methodology, although effective, introduced a level of redundancy into the system.
+
+To optimize this, I experimented with an alternative approach in the index2.ts file. In this modification, I brought in a new role named 'system' within the message structure of the fetchFunction. The 'system' role's content is designed to programmatically determine the user's intent. Preliminary tests have shown that this method is equally robust and reduces unnecessary overhead.
+
+Additionally, I've enhanced the user interface by introducing a more aesthetically pleasing color scheme, ensuring not only functionality but also an friendly user experience.
