+++ package.json
@@ -3,7 +3,7 @@
     "license": "GPL-3.0",
     "packageManager": "yarn@3.6.3",
     "scripts": {
-        "start": "parcel indexcopy.html"
+        "start": "parcel index.html"
     },
     "devDependencies": {
         "buffer": "^5.5.0||^6.0.0",
